refactor(nc-gui,erd): move utils into separate dir
